+++ lib/y.rb
@@ -16,6 +16,7 @@ require_relative "y/rendering"
 require_relative "y/lexxy"
 require_relative "y/lexical_writer"
 require_relative "y/xml_text"
+require_relative "y/doc"
 require_relative "y/tiptap"
 require_relative "y/decoder"